requirements.txt
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246939562/work
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
interval==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.4
mkl-fft==1.2.1
mkl-random==1.1.1
mkl-service==2.3.0
natsort==7.1.1
networkx==2.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
opencv-python==4.5.1.48
Pillow==8.1.0
protobuf==3.15.1
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.17.2
scipy==1.6.1
six @ file:///tmp/build/80754af9/six_1605205313296/work
tensorboardX==2.1
tifffile==2021.2.1
torch==1.7.1
torchvision==0.4.0
tqdm==4.57.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
